refactor: split codegen type runtime emission
